AILLA/src/requirements.txt
asgiref==3.6.0
celery
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==42.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.7
django-allauth==0.51.0
django-cors-headers==3.13.0
django-dynamic-breadcrumbs
django-filter
Django==4.1.10
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
ndg-httpsclient
pyopenssl
pyasn1
gunicorn==20.1.0
idna==3.4
python-magic
oauthlib==3.2.2
openpyxl
pandas
pillow
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.4.0
PyPDF2
pysolr
python3-openid==3.2.0
pytz==2022.7.1
rdflib
requests
requests-oauthlib==1.3.1
sphinx
sphinx_rtd_theme
sqlparse==0.4.4
tox
unidecode
urllib3==1.26.18
validator_collection
watchdog
whitenoise
kazoo

# debugging only
django-debug-toolbar
